+++ lib/src/kernels/transform/mli_krn_tanh_fx.cc
@@ -76,7 +76,7 @@ mli_status mli_krn_tanh_sa8(const mli_tensor* in, mli_tensor* out) {
             &in_params, &out_params, (int)mli_prv_count_elem_num(in));
     out->el_params.sa.zero_point.mem.i16 = out_params.offset;
     out->el_params.sa.scale.mem.i32 = out_params.scale;
-    out->el_params.sa.scale_frac_bits = out_params.shift;
+    out->el_params.sa.scale_frac_bits = (uint8_t)out_params.shift;
 
     return MLI_STATUS_OK;
 }
